api/requirements.txt
fastapi[standard]
mysql-connector-python
py-bcrypt